requirements.txt
fastapi
paho-mqtt==2.1.0